% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bndm_matching}
\alias{bndm_matching}
\title{bndm matching of two strings}
\usage{
bndm_matching(pattern, text)
}
\description{
bndm matching of two strings
}
\examples{
bndm_matching('hallo welt', 'hallo test')
}
