% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bom_matching}
\alias{bom_matching}
\title{bom matching of two strings}
\usage{
bom_matching(pattern, text)
}
\description{
bom matching of two strings
}
\examples{
bom_matching('hallo welt', 'hallo test')
}
