% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cs_shake_v128}
\alias{cs_shake_v128}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{cs_shake_v128}}
\usage{
cs_shake_v128(name, custom_string, input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{cs_shake_v128}
}
\examples{
cs_shake_v128(as.raw("Hallo Welt"),as.raw("Test"),as.raw("Rust"))
}
