% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dot_product_matrix}
\alias{dot_product_matrix}
\title{This function calcluates the dot product of two function}
\usage{
dot_product_matrix(mat_one, mat_two)
}
\description{
This function calcluates the dot product of two function
}
