% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_credit_card_exp}
\alias{get_credit_card_exp}
\title{return a credit card luhn number}
\usage{
get_credit_card_exp()
}
\description{
return a credit card luhn number
}
\examples{
get_credit_card_exp()
}
