% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_credit_card_luhn_number}
\alias{get_credit_card_luhn_number}
\title{return a credit card luhn number}
\usage{
get_credit_card_luhn_number()
}
\description{
return a credit card luhn number
}
\examples{
get_credit_card_luhn_number()
}
