% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_generate_password}
\alias{get_generate_password}
\title{return a fake password}
\usage{
get_generate_password(upper, numeric, special, num)
}
\description{
return a fake password
}
\examples{
get_generate_password(TRUE,TRUE,TRUE,8)
}
