% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_language_abbreviation}
\alias{get_language_abbreviation}
\title{return a language abbreviation}
\usage{
get_language_abbreviation()
}
\description{
return a language abbreviation
}
\examples{
get_language_abbreviation()
}
