% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_language_programming}
\alias{get_language_programming}
\title{return a language programming}
\usage{
get_language_programming()
}
\description{
return a language programming
}
\examples{
get_language_programming()
}
