% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_log_level_appache}
\alias{get_log_level_appache}
\title{return a log level appache}
\usage{
get_log_level_appache()
}
\description{
return a log level appache
}
\examples{
get_log_level_appache()
}
