% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_log_level_general}
\alias{get_log_level_general}
\title{return a log level general}
\usage{
get_log_level_general()
}
\description{
return a log level general
}
\examples{
get_log_level_general()
}
