% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_paragraph}
\alias{get_paragraph}
\title{return a paragraph}
\usage{
get_paragraph(count, sentence_count, word_count, separator)
}
\description{
return a paragraph
}
\examples{
get_paragraph(10,20,10,",")
}
