% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_phone_formatted}
\alias{get_phone_formatted}
\title{return formatted phone number}
\usage{
get_phone_formatted()
}
\description{
return formatted phone number
}
\examples{
get_phone_formatted()
}
