% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_random_number_range}
\alias{get_random_number_range}
\title{get random number in a range}
\usage{
get_random_number_range(min_num, max_num)
}
\description{
get random number in a range
}
\examples{
get_random_number_range(0,1000)
}
