% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sha1}
\alias{get_sha1}
\title{Return sha1 hash of an input string to R.}
\usage{
get_sha1(input)
}
\description{
Return sha1 hash of an input string to R.
}
\examples{
get_sha1("Hallo Welt")
}
