% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sha512}
\alias{get_sha512}
\title{Return sha512 hash of an input string to R.}
\usage{
get_sha512(input)
}
\description{
Return sha512 hash of an input string to R.
}
\examples{
get_sha512("Hallo Welt")
}
