% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_status_codes_general}
\alias{get_status_codes_general}
\title{return a status code}
\usage{
get_status_codes_general()
}
\description{
return a status code
}
\examples{
get_status_codes_general()
}
