% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_status_codes_simple}
\alias{get_status_codes_simple}
\title{return a status code simple}
\usage{
get_status_codes_simple()
}
\description{
return a status code simple
}
\examples{
get_status_codes_simple()
}
