% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_transmission_gear}
\alias{get_transmission_gear}
\title{return a transmission gear}
\usage{
get_transmission_gear()
}
\description{
return a transmission gear
}
\examples{
get_transmission_gear()
}
