% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_chrome}
\alias{get_user_agent_chrome}
\title{return a user agent chrome}
\usage{
get_user_agent_chrome()
}
\description{
return a user agent chrome
}
\examples{
get_user_agent_chrome()
}
