% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_firefox}
\alias{get_user_agent_firefox}
\title{return a user agent firefox}
\usage{
get_user_agent_firefox()
}
\description{
return a user agent firefox
}
\examples{
get_user_agent_firefox()
}
