% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_linux_platform_token}
\alias{get_user_agent_linux_platform_token}
\title{return a user agent platform token}
\usage{
get_user_agent_linux_platform_token()
}
\description{
return a user agent platform token
}
\examples{
get_user_agent_linux_platform_token()
}
