% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_opera}
\alias{get_user_agent_opera}
\title{return a get user agent opera}
\usage{
get_user_agent_opera()
}
\description{
return a get user agent opera
}
\examples{
get_user_agent_opera()
}
