% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_random_platform}
\alias{get_user_agent_random_platform}
\title{return a get user random platform}
\usage{
get_user_agent_random_platform()
}
\description{
return a get user random platform
}
\examples{
get_user_agent_random_platform()
}
