% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_safari}
\alias{get_user_agent_safari}
\title{return a get user agent safari}
\usage{
get_user_agent_safari()
}
\description{
return a get user agent safari
}
\examples{
get_user_agent_safari()
}
