% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_user_agent_windows_platform_token}
\alias{get_user_agent_windows_platform_token}
\title{return a get user windows platform}
\usage{
get_user_agent_windows_platform_token()
}
\description{
return a get user windows platform
}
\examples{
get_user_agent_windows_platform_token()
}
