% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hamming}
\alias{hamming}
\title{Get hamming Distance of two strings}
\usage{
hamming(alpha_str, beta_str)
}
\description{
Get hamming Distance of two strings
}
\examples{
hamming('hallo welt','hallo test')
}
