% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hamming_smid}
\alias{hamming_smid}
\title{hamming_smid matching of two strings}
\usage{
hamming_smid(alpha_str, beta_str)
}
\description{
hamming_smid matching of two strings
}
\examples{
hamming_smid('hallo welt', 'hallo test')
}
