% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{horspool_matching}
\alias{horspool_matching}
\title{horspool matching of two strings}
\usage{
horspool_matching(pattern, text)
}
\description{
horspool matching of two strings
}
\examples{
horspool_matching('hallo welt', 'hallo test')
}
