% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{json_str_to_robj}
\alias{json_str_to_robj}
\title{This function converts an json str to to an RList}
\usage{
json_str_to_robj(input_str)
}
\description{
This function converts an json str to to an RList
}
