% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kangaroo_twelve_xof}
\alias{kangaroo_twelve_xof}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.KangarooTwelveXof.html}{kangaroo_twelve_xof}}
\usage{
kangaroo_twelve_xof(custom_string, input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.KangarooTwelveXof.html}{kangaroo_twelve_xof}
}
\examples{
kangaroo_twelve_xof(as.raw("Hallo Welt"),as.raw("Test"))
}
