% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{keccak_v224}
\alias{keccak_v224}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v224}}
\usage{
keccak_v224(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v224}
}
\examples{
keccak_v224(as.raw("Rust"))
}
