% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{keccak_v256}
\alias{keccak_v256}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v256}}
\usage{
keccak_v256(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v256}
}
\examples{
keccak_v256(as.raw("Rust")) 
}
