% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{keccak_v384}
\alias{keccak_v384}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v384}}
\usage{
keccak_v384(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Keccak.html}{keccak_v384}
}
\examples{
keccak_v384(as.raw("Rust")) 
}
