% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{keccak_v512}
\alias{keccak_v512}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{keccak_v512}}
\usage{
keccak_v512(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{keccak_v512}
}
\examples{
keccak_v512(as.raw("Rust")) 
}
