% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kmac_v128}
\alias{kmac_v128}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Kmac.html}{kmac_v128}}
\usage{
kmac_v128(key, custom_string, input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Kmac.html}{kmac_v128}
}
\examples{
kmac_v128(as.raw("Hallo Welt"),as.raw("Test"),as.raw("Rust")) 
}
