% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kmac_v128_xof}
\alias{kmac_v128_xof}
\title{This is a wrapper for \link{kmac_v128_xof}(/// This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{kmac_v128_xof}}
\usage{
kmac_v128_xof(key, input, custom_string)
}
\description{
This is a wrapper for \link{kmac_v128_xof}(/// This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.CShake.html}{kmac_v128_xof}
}
\examples{
kmac_v128_xof(as.raw("Hallo Welt"),as.raw("Test"),as.raw("Rust")) 
}
