% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kmac_v256_xof}
\alias{kmac_v256_xof}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.KmacXof.html}{kmac_v256_xof}}
\usage{
kmac_v256_xof(key, input, custom_string)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.KmacXof.html}{kmac_v256_xof}
}
\examples{
kmac_v256_xof(as.raw("Hallo Welt"),as.raw("Test"),as.raw("Rust")) 
}
