% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{kmp_matching}
\alias{kmp_matching}
\title{kmp matching of two strings}
\usage{
kmp_matching(pattern, text)
}
\description{
kmp matching of two strings
}
\examples{
kmp_matching('hallo welt', 'hallo test')
}
