% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{levenshtein}
\alias{levenshtein}
\title{Get levenshtein distance of two strings}
\usage{
levenshtein(alpha_str, beta_str)
}
\description{
Get levenshtein distance of two strings
}
\examples{
levenshtein('hallo welt', 'hallo test')
}
