% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{levenshtein_smid}
\alias{levenshtein_smid}
\title{levenshtein_smid matching of two strings}
\usage{
levenshtein_smid(alpha_str, beta_str)
}
\description{
levenshtein_smid matching of two strings
}
\examples{
levenshtein_smid('hallo welt', 'hallo test')
}
