% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{match_regex_string}
\alias{match_regex_string}
\title{This function checks if an input string matches with an regex input}
\usage{
match_regex_string(regex_input, seach_string)
}
\description{
This function checks if an input string matches with an regex input
}
\examples{
match_regex_string("^[0-9]*$","153020")
}
