% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parallel_hash_v256}
\alias{parallel_hash_v256}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.ParallelHash.html}{parallel_hash_v256}}
\usage{
parallel_hash_v256(custom_string, input, block_size)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.ParallelHash.html}{parallel_hash_v256}
}
\examples{
parallel_hash_v256(as.raw("Hallo Welt"),as.raw("Test"),32)  
}
