% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_file_to_string}
\alias{read_file_to_string}
\title{Read a text file out and return a string}
\usage{
read_file_to_string(file_path)
}
\description{
Read a text file out and return a string
}
\examples{
read_file_to_string("test_file1.txt")
}
