% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{read_sheet_by_name}
\alias{read_sheet_by_name}
\title{read excel sheet by name to json}
\usage{
read_sheet_by_name(file_path, sheet_name)
}
\description{
This is a function to read out the content out of a excel sheet to json array
}
\examples{
read_sheet_by_name("example_input.xlsx","Test1")
}
