% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{replace_regex_string}
\alias{replace_regex_string}
\title{This function replaces a regex in an input string}
\usage{
replace_regex_string(input_string, regex_input, replace_str)
}
\description{
This function replaces a regex in an input string
}
\examples{
replace_regex_string("1020 Wien","[^0-9]+","")
}
