% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{resize_image}
\alias{resize_image}
\title{Resize an image}
\usage{
resize_image(path, nwidth, nheight, filter, path_new_image, only_shrink)
}
\description{
The filter option should be Nearest, Triangle, CatmullRom, Gaussian, Lanczos3
The default option for the image Lanczos3
}
\examples{
resize_image('20220824_214140.jpg',500,500,'Nearest','20220824_214140_nearest.jpg',TRUE)
resize_image('20220824_214212.jpg',500,500,'Triangle','20220824_214212_triangle.jpg',TRUE)
resize_image('20220824_214215.jpg',500,500,'CatmullRom','20220824_214215_catmullRom.jpg',TRUE)
resize_image('20220824_214219.jpg',500,500,'Gaussian','20220824_214219_gaussian.jpg',TRUE)
resize_image('20220824_214217.jpg',500,500,'Lanczos3','20220824_214217_lanczos3.jpg',TRUE)
}
