% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R,
%   R/send_custom_http_request.R
\name{send_custom_http_request}
\alias{send_custom_http_request}
\title{return a HTTP Response}
\usage{
send_custom_http_request(
  url,
  method,
  request_headers,
  request_query_params,
  request_form,
  request_body
)

send_custom_http_request(
  url,
  method,
  request_headers,
  request_query_params,
  request_form,
  request_body
)
}
\description{
return a HTTP Response

return a HTTP Response
}
\examples{
send_custom_http_request('https://httpbin.org/get','GET',list(),list(),list(),"")
send_custom_http_request('https://httpbin.org/post','POST',list(),list(),list(),"")
send_custom_http_request('https://httpbin.org/get','GET',list(),list(),list(),"")
send_custom_http_request('https://httpbin.org/post','POST',list(),list(),list(),"")
}
