% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sha3_v224}
\alias{sha3_v224}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Sha3.html}{sha3_v224}}
\usage{
sha3_v224(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Sha3.html}{sha3_v224}
}
\examples{
sha3_v224(as.raw("Hallo Welt"))  
}
