% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sha3_v512}
\alias{sha3_v512}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Sha3.html}{sha3_v512}}
\usage{
sha3_v512(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Sha3.html}{sha3_v512}
}
\examples{
sha3_v512(as.raw("Hallo Welt"))
}
