% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{shake_v128}
\alias{shake_v128}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Shake.html}{shake_v128}}
\usage{
shake_v128(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Shake.html}{shake_v128}
}
\examples{
shake_v128(as.raw("Hallo Welt"))
}
