% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{shake_v256}
\alias{shake_v256}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Shake.html}{shake_v256}}
\usage{
shake_v256(input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.Shake.html}{shake_v256}
}
\examples{
shake_v256(as.raw("Hallo Welt"))
}
