% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{split_regex_string}
\alias{split_regex_string}
\title{This function split an input string to parts}
\usage{
split_regex_string(input_string, regex_input)
}
\description{
This function split an input string to parts
}
\examples{
split_regex_string("a b c d e","[ ]+")
}
