% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tuple_hash_v128}
\alias{tuple_hash_v128}
\title{This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.TupleHash.html}{tuple_hash_v128}}
\usage{
tuple_hash_v128(custom_string, input)
}
\description{
This is a wrapper for \href{https://docs.rs/tiny-keccak/latest/tiny_keccak/struct.TupleHash.html}{tuple_hash_v128}
}
\examples{
tuple_hash_v128(as.raw("Hallo Welt"),as.raw("Rust"))
}
