% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{write_sheet_by_name}
\alias{write_sheet_by_name}
\title{write json to excel}
\usage{
write_sheet_by_name(file_path, sheet_name, json_input_str)
}
\description{
This is a function to write json arrays or json objects to an excel sheet
}
\examples{
write_sheet_by_name('example.xlsx',"Test1",jsonlite::toJSON(mtcars))
}
